ArtificialIntelligence/requirements.txt
fastapi
uvicorn
neo4j
faiss-cpu
sentence-transformers
numpy
